src/nginx-aws-sync/requirements.txt
boto3==1.9.244
PyYAML==5.1.2